Separate prefetch types from prefetch logic
