+++ src/components/routerLink.browser.spec.ts
@@ -4,7 +4,8 @@ import { defineAsyncComponent, h, nextTick, ref } from 'vue'
 import echo from '@/components/echo'
 import { createRoute } from '@/services/createRoute'
 import { createRouter } from '@/services/createRouter'
-import { PrefetchConfig, getPrefetchConfigValue } from '@/types/prefetch'
+import { PrefetchConfig } from '@/types/prefetch'
+import { getPrefetchConfigValue } from '@/utilities/prefetch'
 import { component } from '@/utilities/testHelpers'
 import { visibilityObserverKey } from '@/compositions/useVisibilityObserver'
 import { VisibilityObserver } from '@/services/createVisibilityObserver'
