+++ src/compositions/usePrefetching.ts
@@ -1,7 +1,8 @@
 import { InjectionKey, MaybeRefOrGetter, ref, Ref, toValue, watch } from 'vue'
 import { createUsePropStore } from '@/compositions/usePropStore'
 import { isWithComponent, isWithComponents } from '@/types/createRouteOptions'
-import { getPrefetchOption, PrefetchConfigs, PrefetchStrategy } from '@/types/prefetch'
+import type { PrefetchConfigs, PrefetchStrategy } from '@/types/prefetch'
+import { getPrefetchOption } from '@/utilities/prefetch'
 import { ResolvedRoute } from '@/types/resolved'
 import { isAsyncComponent } from '@/utilities/components'
 import { useVisibilityObserver } from './useVisibilityObserver'
